Move Megatron FSDP tests to mfsdp_v1
